Separate out depth functions in separate class
Rather than a mixin class inherited by ShallowWaterTerm and TracerTerm,
put wd_bathymetry_displacement and get_total_depth in a separate class
that works out the relevant depth expression based on options.
This is because these need to be accesible by diagnostic as well.
Restore get_bnd_functions() on TracerTerm as they were in master.